# perf(compiler): Convert node & field kinds to enums