Runs codegen in the buildscript to generate strongly typed enums for field and node kinds in the parse tree. This is a good idea in general, but also has a slight perf benefit (string matching had to allocate because tree-sitter expected the identifiers to be null-terminated).

## Before
| Stage | Median time (ms) | Min time (ms) | Max time (ms) | Average time (ms) | Standard deviation (ms) |
| --- | --: | --: | --: | --: | --: |
| Init | 0 | 0 | 0 | 0.00 | 0.00 |
| Parse | 125 | 119 | 193 | 126.22 | 7.23 |
| Pass1 | 12 | 11 | 20 | 11.68 | 0.93 |
| Ast | 90 | 86 | 138 | 91.19 | 5.33 |
| Mono | 93 | 87 | 128 | 93.69 | 4.51 |
| Optimizations | 5 | 4 | 6 | 4.79 | 0.47 |
| Codegen | 28 | 26 | 50 | 28.32 | 2.15 |
| TOTAL | 359 | 348 | 503 | 362.62 | 15.41 |

## After
| Stage | Median time (ms) | Min time (ms) | Max time (ms) | Average time (ms) | Standard deviation (ms) |
| --- | --: | --: | --: | --: | --: |
| Init | 0 | 0 | 0 | 0.00 | 0.00 |
| Parse | 126 | 120 | 148 | 126.97 | 3.88 |
| Pass1 | 9 | 8 | 13 | 8.61 | 0.70 |
| Ast | 68 | 65 | 76 | 68.55 | 1.89 |
| Mono | 93 | 88 | 103 | 93.47 | 2.62 |
| Optimizations | 5 | 4 | 7 | 4.94 | 0.44 |
| Codegen | 28 | 26 | 34 | 28.12 | 1.21 |
| TOTAL | 336 | 321 | 366 | 337.06 | 6.90 |